+++ tests/checks/scaling_phase_custom_resource/scaling_phase_custom_resource_test.go
@@ -23,7 +23,8 @@ var (
 	testNamespace        = fmt.Sprintf("%s-ns", testName)
 	rolloutName          = fmt.Sprintf("%s-rollout", testName)
 	serviceName          = fmt.Sprintf("%s-service", testName)
-	httpScaledObjectName = fmt.Sprintf("%s-http-so", testName)
+	interceptorRouteName = fmt.Sprintf("%s-ir", testName)
+	scaledObjectName     = fmt.Sprintf("%s-so", testName)
 	host                 = testName
 	minReplicaCount      = 0
 	maxReplicaCount      = 4
@@ -33,7 +34,8 @@ type templateData struct {
 	TestNamespace        string
 	RolloutName          string
 	ServiceName          string
-	HTTPScaledObjectName string
+	InterceptorRouteName string
+	ScaledObjectName     string
 	Host                 string
 	MinReplicas          int
 	MaxReplicas          int
@@ -122,30 +124,45 @@ spec:
   activeDeadlineSeconds: 600
   backoffLimit: 5
 `
-	httpScaledObjectTemplate = `
-kind: HTTPScaledObject
-apiVersion: http.keda.sh/v1alpha1
+	interceptorRouteTemplate = `
+kind: InterceptorRoute
+apiVersion: http.keda.sh/v1beta1
 metadata:
-  name: {{.HTTPScaledObjectName}}
+  name: {{.InterceptorRouteName}}
   namespace: {{.TestNamespace}}
 spec:
-  hosts:
-  - {{.Host}}
+  target:
+    service: {{.ServiceName}}
+    port: 8080
   scalingMetric:
     requestRate:
       granularity: 1s
       targetValue: 2
       window: 1m
-  scaledownPeriod: 0
+  rules:
+  - hosts:
+    - {{.Host}}
+`
+
+	scaledObjectTemplate = `
+apiVersion: keda.sh/v1alpha1
+kind: ScaledObject
+metadata:
+  name: {{.ScaledObjectName}}
+  namespace: {{.TestNamespace}}
+spec:
+  cooldownPeriod: 0
+  maxReplicaCount: {{ .MaxReplicas }}
+  minReplicaCount: {{ .MinReplicas }}
   scaleTargetRef:
-    name: {{.RolloutName}}
     apiVersion: argoproj.io/v1alpha1
     kind: Rollout
-    service: {{.ServiceName}}
-    port: 8080
-  replicas:
-    min: {{ .MinReplicas }}
-    max: {{ .MaxReplicas }}
+    name: {{.RolloutName}}
+  triggers:
+  - type: external-push
+    metadata:
+      interceptorRoute: {{.InterceptorRouteName}}
+      scalerAddress: keda-add-ons-http-external-scaler.keda:9090
 `
 )
 
@@ -189,14 +206,16 @@ func getTemplateData() (templateData, []Template) {
 			TestNamespace:        testNamespace,
 			RolloutName:          rolloutName,
 			ServiceName:          serviceName,
-			HTTPScaledObjectName: httpScaledObjectName,
+			InterceptorRouteName: interceptorRouteName,
+			ScaledObjectName:     scaledObjectName,
 			Host:                 host,
 			MinReplicas:          minReplicaCount,
 			MaxReplicas:          maxReplicaCount,
 		}, []Template{
 			{Name: "workloadTemplate", Config: workloadTemplate},
 			{Name: "serviceNameTemplate", Config: serviceTemplate},
-			{Name: "httpScaledObjectTemplate", Config: httpScaledObjectTemplate},
+			{Name: "interceptorRouteTemplate", Config: interceptorRouteTemplate},
+			{Name: "scaledObjectTemplate", Config: scaledObjectTemplate},
 		}
 }